🕶 Cleanup validation (#549)
* refactor: move validation into single file

* tests: update tests

* refactor; remove unused

* refactor; removed unused warnings

* tests: Added test for warnings

* refactor: Renaming to be consistent with other warnings
