+++ cognite_toolkit/_cdf_tk/tk_warnings/__init__.py
@@ -9,13 +9,11 @@ from .fileread import (
     CaseTypoWarning,
     DataSetMissingWarning,
     FileReadWarning,
-    MissingRequiredParameter,
+    MissingRequiredParameterWarning,
     NamingConventionWarning,
-    ResourceMissingIdentifier,
-    SnakeCaseWarning,
+    ResourceMissingIdentifierWarning,
     TemplateVariableWarning,
     UnresolvedVariableWarning,
-    UnusedParameter,
     UnusedParameterWarning,
     YAMLFileWarning,
     YAMLFileWithElementWarning,
@@ -37,18 +35,17 @@ __all__ = [
     "ToolkitWarning",
     "GeneralWarning",
     "WarningList",
-    "SnakeCaseWarning",
     "DataSetMissingWarning",
     "TemplateVariableWarning",
     "UnresolvedVariableWarning",
     "UnusedParameter",
     "UnusedParameterWarning",
-    "MissingRequiredParameter",
+    "MissingRequiredParameterWarning",
     "YAMLFileWarning",
     "YAMLFileWithElementWarning",
     "FileReadWarning",
     "NamingConventionWarning",
-    "ResourceMissingIdentifier",
+    "ResourceMissingIdentifierWarning",
     "CaseTypoWarning",
     "ToolkitNotSupportedWarning",
     "UnexpectedFileLocationWarning",
