--- cognite_toolkit/_cdf_tk/validation/__init__.py
@@ -1,3 +1,0 @@
-from .read_yaml import validate_case_raw, validate_data_set_is_set, validate_modules_variables, validate_yaml_config
-
-__all__ = ["validate_modules_variables", "validate_data_set_is_set", "validate_case_raw", "validate_yaml_config"]
